refactor: external module
