+++ tests/gentropy/no_spark/test_external_credentials.py
@@ -8,8 +8,8 @@ from pathlib import Path
 import pytest
 from pydantic import ValidationError
 
-from gentropy.external.hf_hub.credentials import HuggingFaceHubCredentials
-from gentropy.external.wandb.credentials import WandbCredentials
+from gentropy.external.hf_hub import HuggingFaceHubCredentials
+from gentropy.external.wandb import WandbCredentials
 
 DATA_SAMPLES = Path(__file__).parent.parent / "data_samples"
 WANDB_CREDENTIALS_PATH = DATA_SAMPLES / "wandb_credentials.json"
